common names

Return a table of common names

common_names(species_list = NULL, server = getOption("FISHBASE_API",
  "fishbase"), version = get_latest_release(), db = default_db(),
  Language = "English", fields = NULL)

Arguments

species_list
A vector of scientific names (each element as "genus species"). If empty, a table for all fish will be returned.

server
can be set to either "fishbase" or "sealifebase" to switch between databases. NOTE: it is usually easier to leave this as NULL and set the source instead using the environmental variable `FISHBASE_API`, e.g. `Sys.setenv(FISHBASE_API="sealifebase")`.

version
a version string for the database, will default to the latest release. see [get_releases()] for details.

db
the

Language
a string specifying the language for the common name, e.g. "English"

fields
a character vector specifying which fields (columns) should be returned. By default, all available columns recognized by the parser are returned. Mostly for backwards compatibility as users can subset by column later

Value

a data.frame of common names by species queried. If multiple species are queried, The resulting data.frames are concatenated.

Details

Note that there are many common names for a given sci name, so sci_to_common doesn't make sense

Examples

common_names(c("Labroides bicolor",  "Bolbometopon muricatum"))

#> # A tibble: 15 x 4
#>    SpecCode Species                ComName                   Language
#>    <chr>    <chr>                  <chr>                     <chr>   
#>  1 5650     Labroides bicolor      Bicolor cleaner wrasse    English 
#>  2 5650     Labroides bicolor      Bicolor cleanerfish       English 
#>  3 5650     Labroides bicolor      Bicolored cleaner wrasse  English 
#>  4 5650     Labroides bicolor      Bicolour cleaner wrasse   English 
#>  5 5650     Labroides bicolor      Bicoloured cleaner wrasse English 
#>  6 5650     Labroides bicolor      Cleaner wrasse            English 
#>  7 5650     Labroides bicolor      Two-colour cleaner        English 
#>  8 5650     Labroides bicolor      Two-colour cleaner wrasse English 
#>  9 5650     Labroides bicolor      Wrasse                    English 
#> 10 5650     Labroides bicolor      Yellow diesel wrasse      English 
#> 11 5537     Bolbometopon muricatum Bumphead parrotfish       English 
#> 12 5537     Bolbometopon muricatum Double-headed parrotfish  English 
#> 13 5537     Bolbometopon muricatum Green humphead parrotfish English 
#> 14 5537     Bolbometopon muricatum Humphead parrotfish       English 
#> 15 5537     Bolbometopon muricatum Napoleon's fish           English 

# subset by English language names
fish <- common_names("Bolbometopon muricatum")