list_fields

list fields

list_fields(fields = NULL, server = NULL, implemented_only = TRUE)

Arguments

fields
field (column name) to search for

server
base URL to the FishBase API (by default). For SeaLifeBase, use https://fishbase.ropensci.org/sealifebase

implemented_only
by default, only return those tables that have been implemented.

Value

a data frame listing the table names (matching function names in rfishbase) and the matching column names those tables have implemented.

Details

Calling `list_fields()` with no arguments will return the full table of all known fields. Then users can employ standard filter techniques like grep for partial name matching; see examples.

Examples

list_fields("Temp")

#> [90m# A tibble: 2 x 1[39m
#>   table     
#>   [3m[90m<chr>[39m[23m     
#> [90m1[39m occurrence
#> [90m2[39m oxygen    

## Regex matching on full table
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

list_fields() %>% filter(grepl("length", columns, ignore.case = TRUE))

#> [90m# A tibble: 107 x 2[39m
#>    table     columns          
#>    [3m[90m<chr>[39m[23m     [3m[90m<chr>[39m[23m            
#> [90m 1[39m diet      FishLength       
#> [90m 2[39m fecundity LengthFecunMin   
#> [90m 3[39m fecundity LengthTypeFecMin 
#> [90m 4[39m fecundity LengthFecunMax   
#> [90m 5[39m fecundity LengthTypeFecMax 
#> [90m 6[39m fecundity LengthFecunMean  
#> [90m 7[39m fecundity LengthTypeFecMean
#> [90m 8[39m fecundity LengthMin        
#> [90m 9[39m fecundity LengthMax        
#> [90m10[39m fecundity LengthType       
#> [90m# ... with 97 more rows[39m