docs

documentation of tables and fields

docs(table = NULL, server = NULL, ...)

Arguments

table
the table for which the documentation should be displayed. If no table is given, documentation summarizing all available tables is shown.

server
can be set to either "fishbase" or "sealifebase" to switch between databases. NOTE: it is usually easier to leave this as NULL and set the source instead using the environmental variable `FISHBASE_API`, e.g. `Sys.setenv(FISHBASE_API="sealifebase")`.

...
unused; for backwards compatibility only

Value

A data.frame which lists the name of each table (if no table argument is given), along with a description of the table and a URL linking to further information about the table. If a specific table is named in the table argument, then the function will return a data.frame listing all the fields (columns) found in that table, a description of what the field label means, and the units in which the field is measured. These descriptions of the columns are not made available by FishBase.org and must be manually generated and curated by FishBase.org users. At this time, many fields are still missing. Please take a moment to fill in any fields you use in the source table here: https://github.com/ropensci/fishbaseapi/tree/master/docs/docs-sources

Examples

tables <- docs()
# Describe the fecundity table
dplyr::filter(tables, table == "fecundity")$description

#> [1] 
#> 35 Levels:  ...

## See fields in fecundity table
docs("fecundity")

#>    table_name      column_name         type       size
#> 1   fecundity          autoctr Long Integer   4.000000
#> 2   fecundity        StockCode Long Integer   4.000000
#> 3   fecundity         SpecCode Long Integer   4.000000
#> 4   fecundity        MainRefNo Long Integer   4.000000
#> 5   fecundity        SourceRef Long Integer   4.000000
#> 6   fecundity           C_Code         Text   4.000000
#> 7   fecundity           E_CODE Long Integer   4.000000
#> 8   fecundity         Locality         Text 250.000000
#> 9   fecundity     FecundityMin Long Integer   4.000000
#> 10  fecundity        WeightMin       Single   4.000000
#> 11  fecundity   LengthFecunMin       Single   4.000000
#> 12  fecundity LengthTypeFecMin         Text   2.000000
#> 13  fecundity     FecundityMax Long Integer   4.000000
#> 14  fecundity        WeightMax       Single   4.000000
#> 15  fecundity   LengthFecunMax       Single   4.000000
#> 16  fecundity LengthTypeFecMax         Text   2.000000
#> 17  fecundity       FecComment         Text 255.000000
#> 18  fecundity     FecundityRef Long Integer   4.000000
#> 19  fecundity  RelFecundityMin Long Integer   4.000000
#> 20  fecundity      FecunMinRef Long Integer   4.000000
#> 21  fecundity RelFecundityMean Long Integer   4.000000
#> 22  fecundity     FecunMeanRef Long Integer   4.000000
#> 23  fecundity  RelFecundityMax Long Integer   4.000000
#> 24  fecundity      FecunMaxRef Long Integer   4.000000
#> 25  fecundity        LengthMin       Single   4.000000
#> 26  fecundity        LengthMax       Single   4.000000
#> 27  fecundity       LengthType         Text   2.000000
#> 28  fecundity           Number      Integer   2.000000
#> 29  fecundity                a       Single   4.000000
#> 30  fecundity                b       Single   4.000000
#> 31  fecundity               r2       Single   4.000000
#> 32  fecundity              SEa       Single   4.000000
#> 33  fecundity              SEb       Single   4.000000
#> 34  fecundity              SDa       Single   4.000000
#> 35  fecundity              SDb       Single   4.000000
#> 36  fecundity             LCLa       Single   4.000000
#> 37  fecundity             UCLa       Single   4.000000
#> 38  fecundity             LCLb       Single   4.000000
#> 39  fecundity             UCLb       Single   4.000000
#> 40  fecundity   SpawningCycles       Single   4.000000
#> 41  fecundity       SpCycleRef Long Integer   4.000000
#> 42  fecundity         AddInfos         Memo          -
#> 43  fecundity          Entered      Integer   2.000000
#> 44  fecundity      Dateentered    Date/Time   8.000000
#> 45  fecundity         Modified      Integer   2.000000
#> 46  fecundity     Datemodified    Date/Time   8.000000
#> 47  fecundity           Expert      Integer   2.000000
#> 48  fecundity      Datechecked    Date/Time   8.000000
#>                                                                                     description
#> 1                                                        Record number; internal auto counter\n
#> 2                                                                  Code number for internal use
#> 3                                                                  Code number for internal use
#> 4                                      Reference number of main reference used for this record.
#> 5                                                Number of data reference used for this record.
#> 6                                                     3-digit UN numerical country or area code
#> 7                                                    Code number of ecosystem for internal use.
#> 8                                                                Place where sampling was done.
#> 9                                       Absolute fecundity in number of eggs found in a female.
#> 10                                                         Minimum weight of female considered.
#> 11                                                  Length of female to which fecundity refers.
#> 12                                                             Type of length measurement used.
#> 13                                       Absolute fecundity in number of eggs found in a female
#> 14                                                          Maximum weight of female considered
#> 15                                                   Length of female to which fecundity refers
#> 16                                                              Type of length measurement used
#> 17                              Definition of fecundity if not number of eggs found in a female
#> 18                                                     Reference number for absolute fecundity.
#> 19                     Minimum relative fecundity;  no. of mature oocytes per gram total weight
#> 20                                             Reference number for minimum relative fecundity.
#> 21                        Mean relative fecundity;  no. of mature oocytes per gram total weight
#> 22                                                Reference number for mean relative fecundity.
#> 23                     Maximum relative fecundity;  no. of mature oocytes per gram total weight
#> 24                                             Reference number for maximum relative fecundity.
#> 25                         Length of smallest fish considered for fecundity-length relationship
#> 26                          Length of largest fish considered for fecundity-length relationship
#> 27                                                              Type of length measurement used
#> 28                                 Number of specimen(s) used for fecundity-length relationship
#> 29                                   Multiplicative factor (a) of fecundity-length relationship
#> 30                                 Exponent of the fecundity-length relationship (F ;  a x L^b)
#> 31                           Coefficient of determination for the fecundity-length relationship
#> 32             Standard error of the multiplicative factor (a) of fecundity-length relationship
#> 33                     Standard error of the exponent (b) of the fecundity-length relationship.
#> 34         Standard deviation of the multiplicative factor (a) of fecundity-length relationship
#> 35                 Standard deviation of the exponent (b) of the fecundity-length relationship.
#> 36 95% lower confidence limit of the multiplicative factor (a) of fecundity-length relationship
#> 37 95% upper confidence limit of the multiplicative factor (a) of fecundity-length relationship
#> 38              95% lower confidence limit of the exponent (b) of fecundity-length relationship
#> 39              95% upper confidence limit of the exponent (b) of fecundity-length relationship
#> 40              Number of spawning cycles per year;  e.g. 10 in guppies or 0.33 in coelacanths.
#> 41                                                Reference used for number of spawning cycles.
#> 42                                                      Additional information about fecundity.
#> 43                                                   Code number of person who entered the data
#> 44                                                       Date when the record was first entered
#> 45                                                  Code number of person who modified the data
#> 46                                                            Date when the record was modified
#> 47                                                   Code number of person who checked the data
#> 48                                                Date when the record was checked by an expert

## Note: only