synonyms

Check for alternate versions of a scientific name

synonyms(species_list = NULL, server = NULL, ...)

Arguments

species_list
A vector of scientific names (each element as "genus species"). If empty, a table for all fish will be returned.

server
can be set to either "fishbase" or "sealifebase" to switch between databases. NOTE: it is usually easier to leave this as NULL and set the source instead using the environmental variable `FISHBASE_API`, e.g. `Sys.setenv(FISHBASE_API="sealifebase")`.

...
unused; for backwards compatibility only

Value

A table with information about the synonym. Will generally be only a single row if a species name is given. If a FishBase SpecCode is given, all synonyms matching that SpecCode are shown, and the table indicates which one is Valid for FishBase. This may or may not match the valid name for Catalog of Life (Col), also shown in the table. See examples for details.

Details

For further information on fields returned, see: http://www.fishbase.org/manual/english/fishbasethe_synonyms_table.htm

Examples

# Query using a synonym:
synonyms("Callyodon muricatus")

#>               synonym  Status SpecCode SynCode CoL_ID TSN WoRMS_ID ZooBank_ID
#> 1 Callyodon muricatus synonym     5537   59804     NA  NA   312617         NA
#>   TaxonLevel                Species
#> 1    Species Bolbometopon muricatum

 # Check for misspellings or alternate names
 synonyms("Labroides dimidatus") # Species name misspelled

#>               synonym  Status SpecCode SynCode CoL_ID TSN WoRMS_ID ZooBank_ID
#> 1 Labroides dimidatus synonym     5459  149706     NA  NA   402367         NA
#>   TaxonLevel              Species
#> 1    Species Labroides dimidiatus

 # See all synonyms 
 species("Bolbometopon muricatum")

#> [90m# A tibble: 1 x 98[39m
#>   SpecCode Species SpeciesRefNo Author FBname PicPreferredName PicPreferredNam…
#>      [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m          [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m            [3m[90m<chr>[39m[23m           
#> [90m1[39m     [4m5[24m537 Bolbom…         [4m2[24m334 (Vale… Green… Bomur_u4.jpg     [31mNA[39m              
#> [90m# ... with 91 more variables: PicPreferredNameF [3m[90m<chr>[90m[23m, PicPreferredNameJ [3m[90m<chr>[90m[23m,
#> #   FamCode [3m[90m<dbl>[90m[23m, Subfamily [3m[90m<chr>[90m[23m, GenCode [3m[90m<dbl>[90m[23m, SubGenCode [3m[90m<dbl>[90m[23m,
#> #   BodyShapeI [3m[90m<chr>[90m[23m, Source [3m[90m<chr>[90m[23m, AuthorRef [3m[90m<lgl>[90m[23m, Remark [3m[90m<chr>[90m[23m,
#> #   TaxIssue [3m[90m<dbl>[90m[23m, Fresh [3m[90m<dbl>[90m[23m, Brack [3m[90m<dbl>[90m[23m, Saltwater [3m[90m<dbl>[90m[23m,
#> #   DemersPelag [3m[90m<chr>[90m[23m, AnaCat [3m[90m<chr>[90m[23m, MigratRef [3m[90m<dbl>[90m[23m, DepthRangeShallow [3m[90m<dbl>[90m[23m,
#> #   DepthRangeDeep [3m[90m<dbl>[90m[23m, DepthRangeRef [3m[90m<dbl>[90m[23m, DepthRangeComShallow [3m[90m<dbl>[90m[23m,
#> #   DepthRangeComDeep [3m[90m<dbl>[90m[23m, DepthComRef [3m[90m<dbl>[90m[23m, LongevityWild [3m[90m<dbl>[90m[23m,
#> #   LongevityWildRef [3m[90m<dbl>[90m[23m, LongevityCaptive [3m[90m<dbl>[90m[23m, LongevityCapRef [3m[90m<dbl>[90m[23m,
#> #   Vulnerability [3m[90m<dbl>[90m[23m, Length [3m[90m<dbl>[90m[23m, LTypeMaxM [3m[90m<chr>[90m[23m, LengthFemale [3m[90m<dbl>[90m[23m,
#> #   LTypeMaxF [3m[90m<chr>[90m[23m, MaxLengthRef [3m[90m<dbl>[90m[23m, CommonLength [3m[90m<dbl>[90m[23m, LTypeComM [3m[90m<chr>[90m[23m,
#> #   CommonLengthF [3m[90m<dbl>[90m[23m, LTypeComF [3m[90m<chr>[90m[23m, CommonLengthRef [3m[90m<dbl>[90m[23m, Weight [3m[90m<dbl>[90m[23m,
#> #   WeightFemale [3m[90m<dbl>[90m[23m, MaxWeightRef [3m[90m<dbl>[90m[23m, Pic [3m[90m<chr>[90m[23m, PictureFemale [3m[90m<chr>[90m[23m,
#> #   LarvaPic [3m[90m<chr>[90m[23m, EggPic [3m[90m<lgl>[90m[23m, ImportanceRef [3m[90m<dbl>[90m[23m, Importance [3m[90m<chr>[90m[23m,
#> #   PriceCateg [3m[90m<chr>[90m[23m, PriceReliability [3m[90m<chr>[90m[23m, Remarks7 [3m[90m<chr>[90m[23m,
#> #   LandingStatistics [3m[90m<chr>[90m[23m, Landings [3m[90m<chr>[90m[23m, MainCatchingMethod [3m[90m<chr>[90m[23m,
#> #   II [3m[90m<chr>[90m[23m, MSeines [3m[90m<dbl>[90m[23m, MGillnets [3m[90m<dbl>[90m[23m, MCastnets [3m[90m<dbl>[90m[23m, MTraps [3m[90m<dbl>[90m[23m,
#> #   MSpears [3m[90m<dbl>[90m[23m, MTrawls [3m[90m<dbl>[90m[23m, MDredges [3m[90m<dbl>[90m[23m, MLiftnets [3m[90m<dbl>[90m[23m,
#> #   MHooksLines [3m[90m<dbl>[90m[23m, MOther [3m[90m<dbl>[90m[23m, UsedforAquaculture [3m[90m<chr>[90m[23m, LifeCycle [3m[90m<lgl>[90m[23m,
#> #   AquacultureRef [3m[90m<dbl>[90m[23m, UsedasBait [3m[90m<chr>[90m[23m, BaitRef [3m[90m<dbl>[90m[23m, Aquarium [3m[90m<chr>[90m[23m,
#> #   AquariumFishII [3m[90m<chr>[90m[23m, AquariumRef [3m[90m<dbl>[90m[23m, GameFish [3m[90m<dbl>[90m[23m, GameRef [3m[90m<dbl>[90m[23m,
#> #   Dangerous [3m[90m<chr>[90m[23m, DangerousRef [3m[90m<dbl>[90m[23m, Electrogenic [3m[90m<chr>[90m[23m, ElectroRef [3m[90m<dbl>[90m[23m,
#> #   Complete [3m[90m<lgl>[90m[23m, GoogleImage [3m[90m<dbl>[90m[23m, Comments [3m[90m<chr>[90m[23m, Profile [3m[90m<lgl>[90m[23m,
#> #   PD50 [3m[90m<dbl>[90m[23m, Emblematic [3m[90m<dbl>[90m[23m, Entered [3m[90m<dbl>[90m[23m, DateEntered [3m[90m<dttm>[90m[23m,
#> #   Modified [3m[90m<dbl>[90m[23m, DateModified [3m[90m<dttm>[90m[23m, Expert [3m[90m<dbl>[90m[23m, DateChecked [3m[90m<dttm>[90m[23m,
#> #   TS [3m[90m<lgl>[90m[23m[39m